Fixed the main part of Taiga issue #129 (GitHub Issue #19).
Removed comment locations were not being spliced out of the comment's list of
locations correctly (all locations after the one to remove were removed from the
list, not just the one to be removed). There is one more related issue with
badge counts not being updated that has yet to be fixed.